+++ internal/api/types/api_server.pb.gw.go
@@ -92,7 +92,7 @@ func request_API_CheckWithdrawal_0(ctx context.Context, marshaler runtime.Marsha
 	if !ok {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "tx_nonce")
 	}
-	protoReq.TxNonce, err = runtime.Uint64(val)
+	protoReq.TxNonce, err = runtime.Uint32(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "tx_nonce", err)
 	}
@@ -126,7 +126,7 @@ func local_request_API_CheckWithdrawal_0(ctx context.Context, marshaler runtime.
 	if !ok {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "tx_nonce")
 	}
-	protoReq.TxNonce, err = runtime.Uint64(val)
+	protoReq.TxNonce, err = runtime.Uint32(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "tx_nonce", err)
 	}
